Move `DispatchQueue` combine helpers to a separate file
