+++ EssentialApp/EssentialApp/CombineHelpers.swift
@@ -81,47 +81,3 @@ extension Publisher {
         self.catch { _ in fallbackPublisher() }.eraseToAnyPublisher()
     }
 }
-
-extension Publisher {
-    func dispatchOnMainQueue() -> AnyPublisher<Output, Failure> {
-        receive(on: DispatchQueue.mainQueueScheduler).eraseToAnyPublisher()
-    }
-}
-
-extension DispatchQueue {
-    static var mainQueueScheduler = MainQueueScheduler.shared
-
-    struct MainQueueScheduler: Scheduler {
-        typealias SchedulerTimeType = DispatchQueue.SchedulerTimeType
-        typealias SchedulerOptions = DispatchQueue.SchedulerOptions
-
-        var now = DispatchQueue.main.now
-        var minimumTolerance = DispatchQueue.main.minimumTolerance
-
-        public static let shared = Self()
-
-        private static let key = DispatchSpecificKey<UInt8>()
-        private static let value = UInt8.max
-
-        private init() {
-            DispatchQueue.main.setSpecific(key: Self.key, value: Self.value)
-        }
-
-        private func isMainQueue() -> Bool {
-            return DispatchQueue.getSpecific(key: Self.key) == Self.value
-        }
-
-        func schedule(options: DispatchQueue.SchedulerOptions?, _ action: @escaping () -> Void) {
-            guard isMainQueue() else { return DispatchQueue.main.async { action() } }
-            action()
-        }
-
-        func schedule(after date: DispatchQueue.SchedulerTimeType, tolerance: DispatchQueue.SchedulerTimeType.Stride, options: DispatchQueue.SchedulerOptions?, _ action: @escaping () -> Void) {
-            DispatchQueue.main.schedule(after: date, tolerance: tolerance, options: options, action)
-        }
-
-        func schedule(after date: DispatchQueue.SchedulerTimeType, interval: DispatchQueue.SchedulerTimeType.Stride, tolerance: DispatchQueue.SchedulerTimeType.Stride, options: DispatchQueue.SchedulerOptions?, _ action: @escaping () -> Void) -> Cancellable {
-            return DispatchQueue.main.schedule(after: date, interval: interval, tolerance: tolerance, options: options, action)
-        }
-    }
-}
